// apps/admin/src/lib/api/generated/models/connectorResponseOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.51
 */
import type { JsonValue } from './jsonValue';

export interface ConnectorResponseOut {
  data: JsonValue | null;
  sampled: boolean;
}